fix(assignment): add label management on create and edit page
extract label management as separate component
